requirements/base.txt
Django==5.0.4
python-environ==0.4.54
psycopg2==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
django-taggit==5.0.1
django-filter==24.2